backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
openai>=1.3.5
cohere>=4.9.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.3
pytest>=7.4.3
pytest-asyncio>=0.23.2
httpx>=0.25.0